CDAP-4856 Renamed TeraDataSource to DBSource and fixed all the tests
