+++ database-plugins/src/test/java/co/cask/hydrator/plugin/DatabasePluginTestBase.java
@@ -35,10 +35,9 @@ import co.cask.cdap.test.TestBase;
 import co.cask.cdap.test.TestConfiguration;
 import co.cask.hydrator.plugin.db.batch.sink.DBSink;
 import co.cask.hydrator.plugin.db.batch.sink.ETLDBOutputFormat;
-import co.cask.hydrator.plugin.db.batch.source.DBSource;
 import co.cask.hydrator.plugin.db.batch.source.DataDrivenETLDBInputFormat;
 import co.cask.hydrator.plugin.db.batch.source.ETLDBInputFormat;
-import co.cask.hydrator.plugin.teradata.batch.source.TeradataSource;
+import co.cask.hydrator.plugin.db.batch.source.DBSource;
 import co.cask.hydrator.plugin.test.TableSink;
 import co.cask.hydrator.plugin.test.TableSource;
 import com.google.common.base.Charsets;
@@ -110,7 +109,7 @@ public class DatabasePluginTestBase extends TestBase {
 
     addPluginArtifact(Id.Artifact.from(Id.Namespace.DEFAULT, "database-plugins", "1.0.0"), APP_ARTIFACT_ID,
                       DBSource.class, DBSink.class, DBRecord.class, ETLDBInputFormat.class, ETLDBOutputFormat.class,
-                      TeradataSource.class, DataDrivenETLDBInputFormat.class, DBRecord.class);
+                      DataDrivenETLDBInputFormat.class, DBRecord.class);
 
     // add hypersql 3rd party plugin
     PluginClass hypersql = new PluginClass("jdbc", "hypersql", "hypersql jdbc driver", JDBCDriver.class.getName(),
